move static method to processor instead of equiv generator
